Updated to Zenilib v4.0.0 & Started adding in the GUI Code.  Currently Non-Functional however within a day or two a version should be uploaded which is functional

+++ include/Zeni/Quaternion.hxx
@@ -1,36 +1,26 @@
-/* This file is part of the Zenipex Library.
-* Copyleft (C) 2011 Mitchell Keith Bloch a.k.a. bazald
-*
-* The Zenipex Library is free software; you can redistribute it and/or 
-* modify it under the terms of the GNU General Public License as 
-* published by the Free Software Foundation; either version 2 of the 
-* License, or (at your option) any later version.
-*
-* The Zenipex Library is distributed in the hope that it will be useful, 
-* but WITHOUT ANY WARRANTY; without even the implied warranty of 
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
-* General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License 
-* along with the Zenipex Library; if not, write to the Free Software 
-* Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 
-* 02110-1301 USA.
-*
-* As a special exception, you may use this file as part of a free software
-* library without restriction.  Specifically, if other files instantiate
-* templates or use macros or inline functions from this file, or you compile
-* this file and link it with other files to produce an executable, this
-* file does not by itself cause the resulting executable to be covered by
-* the GNU General Public License.  This exception does not however
-* invalidate any other reasons why the executable file might be covered by
-* the GNU General Public License.
-*/
+/* This file is part of the Zenipex Library (zenilib).
+ * Copyright (C) 2011 Mitchell Keith Bloch (bazald).
+ *
+ * zenilib is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * zenilib is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with zenilib.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 #ifndef ZENI_QUATERNION_HXX
 #define ZENI_QUATERNION_HXX
 
 // HXXed below
 #include <Zeni/Coordinate.h>
+#include <Zeni/Matrix4f.h>
 #include <Zeni/Vector3f.h>
 
 #include <Zeni/Quaternion.h>
@@ -131,7 +121,7 @@ namespace Zeni {
   }
 
   float Quaternion::magnitude() const {
-    return sqrt(magnitude2());
+    return float(sqrt(magnitude2()));
   }
 
   Quaternion Quaternion::conjugate() const {
@@ -197,8 +187,8 @@ namespace Zeni {
     else if(q.time > 1.0f)
       q.time = 1.0f;
 
-    const float angle = 2.0f * acos(q.time);
-    float s = sqrt(1.0f - q.time * q.time);
+    const float angle = 2.0f * float(acos(q.time));
+    float s = float(sqrt(1.0f - q.time * q.time));
 
     if(fabs(s) < 0.001f)
       return std::make_pair(q.space, angle);
@@ -251,6 +241,7 @@ namespace Zeni {
 }
 
 #include <Zeni/Coordinate.hxx>
+#include <Zeni/Matrix4f.hxx>
 #include <Zeni/Vector3f.hxx>
 
 #endif
